# perf: reduce charts initialization time up to ~2x