--- a/packages/vaadin-charts/src/vaadin-chart.js
+++ b/packages/vaadin-charts/src/vaadin-chart.js
@@ -981,10 +981,11 @@ class ChartElement extends ElementMixin(ThemableMixin(PolymerElement)) {
         this.__setYAxisProps(yAxes, unit, { max: valueMax });
       }
 
-      const seriesConfiguration = this.__updateOrAddSeriesInstance(seriesElement.options, idxOnChildList);
+      const seriesConfiguration = this.__updateOrAddSeriesInstance(seriesElement.options, idxOnChildList, false);
 
       seriesElement.setSeries(seriesConfiguration);
     }
+
     this.__removeAxisIfEmpty();
   }
 
@@ -1454,11 +1455,11 @@ class ChartElement extends ElementMixin(ThemableMixin(PolymerElement)) {
   }
 
   /** @private */
-  __updateOrAddSeriesInstance(seriesOptions, position) {
+  __updateOrAddSeriesInstance(seriesOptions, position, redraw) {
     if (this.configuration.series[position]) {
-      this.configuration.series[position].update(seriesOptions);
+      this.configuration.series[position].update(seriesOptions, redraw);
     } else {
-      this.configuration.addSeries(seriesOptions);
+      this.configuration.addSeries(seriesOptions, redraw);
     }
     return this.configuration.series[position];
   }