--- a/packages/vaadin-charts/src/vaadin-chart.js
+++ b/packages/vaadin-charts/src/vaadin-chart.js
@@ -985,7 +985,6 @@ class ChartElement extends ElementMixin(ThemableMixin(PolymerElement)) {
 
       seriesElement.setSeries(seriesConfiguration);
     }
-
     this.__removeAxisIfEmpty();
   }