--- a/packages/vaadin-charts/src/vaadin-chart.js
+++ b/packages/vaadin-charts/src/vaadin-chart.js
@@ -986,6 +986,8 @@ class ChartElement extends ElementMixin(ThemableMixin(PolymerElement)) {
       seriesElement.setSeries(seriesConfiguration);
     }
     this.__removeAxisIfEmpty();
+
+    this.configuration.redraw();
   }
 
   /** @private */