--- a/packages/vaadin-charts/test/chart-element.test.js
+++ b/packages/vaadin-charts/test/chart-element.test.js
@@ -1,6 +1,6 @@
 import sinon from 'sinon';
 import { expect } from '@esm-bundle/chai';
-import { aTimeout, fixtureSync, oneEvent } from '@vaadin/testing-helpers';
+import { aTimeout, fixtureSync, nextFrame, oneEvent } from '@vaadin/testing-helpers';
 import '../vaadin-chart.js';
 
 describe('vaadin-chart', () => {
@@ -390,6 +390,28 @@ describe('vaadin-chart', () => {
       const scrollWidth = document.documentElement.scrollWidth;
       document.dir = 'rtl';
       expect(scrollWidth).to.be.equal(document.documentElement.scrollWidth);
+    });
+  });
+
+  describe('performance', () => {
+    let chart, redrawSpy;
+
+    beforeEach(async () => {
+      chart = fixtureSync(`<vaadin-chart></vaadin-chart>`);
+      await oneEvent(chart, 'chart-load');
+
+      redrawSpy = sinon.spy(chart.configuration, 'redraw');
+    });
+
+    describe('adding a series', () => {
+      it.only('should redraw the chart only once', async () => {
+        const series = fixtureSync(`<vaadin-chart-series values="[1, 2, 3, 4]"></vaadin-chart-series>`);
+
+        chart.appendChild(series);
+        await nextFrame();
+
+        expect(redrawSpy.calledOnce).to.be.true;
+      });
     });
   });
 });