--- a/packages/vaadin-charts/test/chart-element.test.js
+++ b/packages/vaadin-charts/test/chart-element.test.js
@@ -424,7 +424,7 @@ describe('vaadin-chart', () => {
         redrawSpy.resetHistory();
       });
 
-      it('should redraw the chart 3 times', async () => {
+      it('should redraw the chart only 3 times', async () => {
         const series = fixtureSync(`<vaadin-chart-series values="[1, 2, 3, 4]"></vaadin-chart-series>`);
 
         chart.replaceChild(series, chart.firstElementChild);