## Description

This PR reduces the initialization time of charts up to ~2x by replacing a `redraw` call in a loop (by series) with a single `redraw` at the end of the loop.

~~This PR also is assumed to improve the situation described in the https://github.com/vaadin/flow-components/issues/937 issue.~~

Part of  #227

## Benchmarks

**Machine:** Mac Mini M1 16gb.

**Browser:** Chrome 92.

**Screen size:** 700x700.

### 1. Basic case (10 series)

- Each series represents 20 values

```html
<div id="chart"></div>

<script type="module">
  import { render, html } from 'lit';
  import '@vaadin/vaadin-charts';

  const series = [...new Array(10)].map((_, i) => {
    return {
      values: [...new Array(20)].map(() => i)
    }
  });

  render(html`
    <vaadin-chart>
      ${series.map((series) =>
        html`
          <vaadin-chart-series .values="${series.values}"></vaadin-chart-series>
        `
      )}
    </vaadin-chart>
  `, document.querySelector('#chart'))
</script>
```

**Master branch:** 465 ms ([screenshot](https://user-images.githubusercontent.com/5039436/129702222-0c2f143a-e97a-4176-9c9f-ee92333b9a53.png))

**Current branch:** 145 ms ([screenshot](https://user-images.githubusercontent.com/5039436/129702368-3858c08c-c436-4f08-bd9f-9ea7749eb9da.png))

**Total:** up to ~3.2x

### 2. More advanced case (10 series)

- Each series represents 20 values
- Each series has a unit
- Each series has a title

```html
<div id="chart"></div>

<script type="module">
  import { render, html } from 'lit';
  import '@vaadin/vaadin-charts';

  const series = [...new Array(10)].map((_, i) => {
    return {
      unit: `Unit ${i}`,
      title: `Title ${i}`,
      values: [...new Array(20)].map(() => i)
    }
  });

  render(html`
    <vaadin-chart>
      ${series.map((series) =>
        html`
          <vaadin-chart-series
            .unit="${series.unit}"
            .title="${series.title}"
            .values="${series.values}"
          ></vaadin-chart-series>
        `
      )}
    </vaadin-chart>
  `, document.querySelector('#chart'))
</script>
```

**Master branch:** 4.53 s ([screenshot](https://user-images.githubusercontent.com/5039436/129702064-d09aeccb-a880-4730-bbf0-6abed9e2e75b.png))

**Current branch:** 2.84 s ([screenshot](https://user-images.githubusercontent.com/5039436/129701968-b632e17a-473d-4282-a371-c37fb3a80304.png))

**Total:** up to ~1.6x

## Type of change

- [x] Internal change

## Checklist

- [x] I have read the contribution guide: https://vaadin.com/docs-beta/latest/guide/contributing/overview/
- [x] I have added a description following the guideline.
- [x] The issue is created in the corresponding repository and I have referenced it.
- [x] I have added tests to ensure my change is effective and works as intended.
- [x] New and existing tests are passing locally with my change.
- [x] I have performed self-review and corrected misspellings.