@vursen 2021-08-17
While this PR improves the state of performance for charts, it is still possible to optimize the initialization time even more by disabling re-drawing in `vaadin-chart-series`'s observers at the time of adding a series and calling a single re-drawing at the end.